requirements.txt
diffusers>=0.33.0
transformers
accelerate
huggingface_hub
sentencepiece
torch
numpy
Pillow
